feat: Pull out styles/html into separate PR #1730
