+++ elements/pfe-icon/src/pfe-icon.js
@@ -1,3 +1,6 @@
+// Import polyfills: window.CustomEvent
+import "./polyfills--pfe-icon.js";
+
 import PFElement from "../../pfelement/dist/pfelement.js";
 import PfeIconSet from "./icon-set.js";
 import { addBuiltIns } from "./builtin-icon-sets.js";
