+++ elements/pfelement/src/pfelement.js
@@ -475,10 +475,7 @@ class PFElement extends HTMLElement {
    */
   render() {
     this.shadowRoot.innerHTML = "";
-    this.template.innerHTML = "";
-
-    if (this.styles) this.template.innerHTML += this.styles;
-    if (this.html) this.template.innerHTML += this.html;
+    this.template.innerHTML = this.html;
 
     if (window.ShadyCSS) {
       window.ShadyCSS.prepareTemplate(this.template, this.tag);
